Default to DelegatingSecurityContextRepository
Closes gh-12023
Closes gh-12049
